REFACTORING: improving the overall code base

+++ Controllers/UsuariosController.cs
@@ -1,10 +1,7 @@
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
 using Microsoft.AspNetCore.Mvc;
 using ProjetoMVC.Models;
 using ProjetoMVC.context;
+using System.Text.RegularExpressions;
 
 namespace ProjetoMVC.Controllers
 {
@@ -27,23 +24,45 @@ namespace ProjetoMVC.Controllers
             return View();
         }
 
+        //As soon as it validates the fields, the method creates a new user in the database.
         [HttpPost]
-        public IActionResult Criar(Usuarios usuario)
+        public async Task<IActionResult> Criar(Usuarios usuario)
         {
+            if (string.IsNullOrEmpty(usuario.Nome) || string.IsNullOrEmpty(usuario.Email) || string.IsNullOrEmpty(usuario.Telefone) ||
+                string.IsNullOrEmpty(usuario.Senha))
+            {
+                throw new ArgumentException("Nome, Email, Telefone e Senha são campos obrigatórios, verifique se não deixou algum campo vazio.");
+            }
+            else if (!Regex.IsMatch(usuario.Email, @"^[\w-\.]+@([\w-]+\.)+[\w-]{2,4}$"))
+            {
+                throw new ArgumentException("Email", "Email está em um formato inválido");
+            }
+
+            if (!usuario.Telefone.All(char.IsDigit))
+            {
+                throw new ArgumentException("Telefone", "O telefone deve conter apenas dígitos.");
+            }
+
+            if (usuario.Senha != usuario.ConfirmarSenha)
+            {
+                throw new ArgumentException("Senhas", "As senhas não são iguais, verifique se digitou corretamente");
+            }
+
             if (ModelState.IsValid)
             {
-                _context.Usuarios.Add(usuario);
-                _context.SaveChanges();
+                await _context.Usuarios.AddAsync(usuario);
+                await _context.SaveChangesAsync();
 
                 return RedirectToAction(nameof(Index));
             }
             return View(usuario);
         }
 
+        //This method is for when accesses the application URL with the id wants.
         [HttpGet]
-        public IActionResult Editar(int id)
+        public async Task<IActionResult> Editar(int id)
         {
-            var usuario = _context.Usuarios.Find(id);
+            var usuario = await _context.Usuarios.FindAsync(id);
             if (usuario == null)
             {
                 return NotFound();
@@ -52,26 +71,51 @@ namespace ProjetoMVC.Controllers
             return View(usuario);
         }
 
+        //This method is for when you want to edit a user's data.
         [HttpPost]
-        public IActionResult Editar(Usuarios usuario)
+        public async Task<IActionResult> Editar(Usuarios usuario)
         {
-            var usuarioBanco = _context.Usuarios.Find(usuario.Id);
+            var usuarioBanco = await _context.Usuarios.FindAsync(usuario.Id);
+
+            if (usuario == null)
+            {
+                return NotFound();
+            }
+
+            if (!Regex.IsMatch(usuario.Email, @"^[\w-\.]+@([\w-]+\.)+[\w-]{2,4}$"))
+            {
+                throw new ArgumentException("Email", "Email está em formato inválido");
+            }
+
+            if (!usuario.Telefone.All(char.IsDigit))
+            {
+                throw new ArgumentException("Telefone", "O telefone deve conter apenas dígitos");
+            }
+
+            // to fix later
+            // if (usuario.Senha != usuario.ConfirmarSenha)
+            // {
+            //     throw new ArgumentException("Senhas", "As senhas não são iguais, verifique se digitou corretamente");
+            // }
 
             usuarioBanco.Nome = usuario.Nome;
             usuarioBanco.Email = usuario.Email;
             usuarioBanco.Telefone = usuario.Telefone;
             usuarioBanco.Senha = usuario.Senha;
+            usuarioBanco.ConfirmarSenha = usuario.ConfirmarSenha;
+
 
             _context.Usuarios.Update(usuarioBanco);
-            _context.SaveChanges();
+            await _context.SaveChangesAsync();
 
             return RedirectToAction(nameof(Index));
         }
 
+        //Search user by id for after delete option.
         [HttpGet]
-        public IActionResult Deletar(int id)
+        public async Task<IActionResult> Deletar(int id)
         {
-            var usuario = _context.Usuarios.Find(id);
+            var usuario = await _context.Usuarios.FindAsync(id);
 
             if (usuario == null)
             {
@@ -81,13 +125,17 @@ namespace ProjetoMVC.Controllers
             return View(usuario);
         }
 
+        //This method delete a user in database.
         [HttpPost]
-        public IActionResult Deletar(Usuarios usuarios)
+        public async Task<IActionResult> Deletar(Usuarios usuarios)
         {
-            var usuarioBanco = _context.Usuarios.Find(usuarios.Id);
+            var usuarioBanco = await _context.Usuarios.FindAsync(usuarios.Id);
 
-            _context.Usuarios.Remove(usuarioBanco);
-            _context.SaveChanges();
+            if (usuarioBanco != null)
+            {
+                _context.Usuarios.Remove(usuarioBanco);
+                await _context.SaveChangesAsync();
+            }
 
             return RedirectToAction(nameof(Index));
         }
